Add book sections
This is a first, high-level pass at adding the documents exported from
Confluence as book chapters.

For now the book imitates the structure of the original guide on
Confluence. While we will want to keep the top-level chapters, we may
want to revisit the structure of the subsections depending on how it
renders in the book.

There need to be new lines at the end or beginning of markdown files.
Otherwise the content is not correctly parsed after contatenation into a
single source file.

Images and links don't display correctly yet. We will need to update
some references.