Another Markdown formatting pass
I opted to remove the space from the top of each of the markdown files,
and add it to `book.md`, I feel like it will be easier to
remember/manage that necessary space in book.md than it will be in the
individual files.

I fixed a few broken links, and generally cleaned things up.