Update Conventions on our latest practices for Kubernetes manifests (#9)
We used to recommend having a separate repository for Kubernetes
manifests. On our new buildouts, we are no longer keeping a separate
repository for manifests, and instead, we keep a folder with Helm Charts
values files to build Kubernetes manifests, and that's what we want to
recommend in the guide going forward.